add transfer transaction model

+++ src/api/models/timestamp.rs
@@ -4,6 +4,6 @@ use serde::Serialize;
 #[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
 pub struct Timestamp {
     pub epoch: u32,
-    pub unix: u32,
+    pub unix: u64,
     pub human: String,
 }
